refactor: move ProfileImage and ProfileIcon to ui/profile
Moved profile-related UI components to dedicated ui/profile/ folder:
- ProfileImage and ProfileIcon are pure UI components used across multiple domains
- Previously located in card/portfolio/components/
- Now in ui/profile/ for better discoverability and reusability

File moves:
- card/portfolio/components/ProfileImage.tsx → ui/profile/ProfileImage.tsx
- card/portfolio/components/ProfileIcon.tsx → ui/profile/ProfileIcon.tsx

Updated imports in 8 files:
- ProfileImage: TeamSidebar, BusinessCard, ProjectTeamSection,
  ProjectCardMetadata, PortfolioCard
- ProfileIcon: Portfolio.tsx, Team.tsx, ProjectPage.tsx

Benefits:
- Clear separation: profile UI components now in ui/ alongside other reusable components
- Better discoverability: developers can easily find profile-related UI in ui/profile/
- Consistent structure: follows the pattern of ui/button/, ui/badge/, etc.

Verified with pnpm lint and pnpm build.
