+++ src/app/components/section/project/ProjectTeamSection.tsx
@@ -1,7 +1,7 @@
 import { Label, Label2 } from '@/app/components/ui/text/text';
 import type { ProjectDetailViewModel } from '@/services/project/types';
 import Link from 'next/link';
-import ProfileImage from '@/app/components/card/portfolio/components/ProfileImage';
+import ProfileImage from '@/app/components/ui/profile/ProfileImage';
 
 interface TeamMemberItemProps {
   member: ProjectDetailViewModel['team'][number];
